// src/api/models/CreateLeadDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateLeadDto = {
    email: string | null;
    phoneNumber: string | null;
    site: string | null;
    name: string | null;
    comment: string | null;
    companyTypeId: string;
};